pyproject.toml
[project]
name = "agentframework"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.70.0",
    "litellm>=1.83.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.0.1",
    "rich>=14.3.3",
]


[[tool.uv.index]]
name = "tuna"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
